LayoutTests/media/picture-in-picture/picture-in-picture-api-enter-pip-4-expected.txt

This tests that request Picture-in-Picture resolves on user click.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplaythrough)
END OF TEST
